- remove the i2c, serial, matrix and split_util files - regenerate some files from the templates
