+++ keyboards/pinky/pinky.h
@@ -1,4 +1,22 @@
-#pragma once
+/* Copyright 2018 'Masayuki Sunahara'
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ */
+#ifndef PINKY_H
+#define PINKY_H
+
+#include "quantum.h"
 
 #ifdef KEYBOARD_pinky_3
   #include "3.h"
@@ -6,4 +24,6 @@
 
 #ifdef KEYBOARD_pinky_4
   #include "4.h"
+#endif
+
 #endif
